requirements.txt
async_timeout==3.0.1
discord.py==2.2.2
matplotlib==3.6.2
numpy==1.21.4
oauth2client==4.1.3
opencv_python==4.5.4.60
requests==2.26.0
selenium==4.8.2
webdriver_manager==3.8.5
